17: Further 18x speed up by forcing evaluation
By adding bang patterns to the critical loop we get an 18x speed up.
From 12 seconds down to 0.5. I assume this makes haskell able to
optimize this down to a simple for loop. However, the naive Rust
solution with a simple for loop takes less than a millisecond..

Haskell performance seems really unintuitive to me.